added common switcher component, and moved add component to left
