+++ packages/builder/src/components/userInterface/ItemTab/Tab.svelte
@@ -3,7 +3,6 @@
   const dispatch = createEventDispatcher()
 
   import Item from "./Item.svelte"
-  import { store } from "builderStore"
   export let list
 
   let category = list
