+++ packages/builder/src/pages/[application]/frontend/_layout.svelte
@@ -2,18 +2,17 @@
   import { store, backendUiStore } from "builderStore"
   import { goto } from "@sveltech/routify"
   import { onMount } from "svelte"
-  import ComponentsHierarchy from "components/userInterface/ComponentsHierarchy.svelte"
   import ComponentsHierarchyChildren from "components/userInterface/ComponentsHierarchyChildren.svelte"
-  import PageLayout from "components/userInterface/PageLayout.svelte"
-  import PagesList from "components/userInterface/PagesList.svelte"
   import IconButton from "components/common/IconButton.svelte"
-  import NewScreen from "components/userInterface/NewScreen.svelte"
   import CurrentItemPreview from "components/userInterface/AppPreview"
   import PageView from "components/userInterface/PageView.svelte"
-  import ComponentsPaneSwitcher from "components/userInterface/ComponentsPaneSwitcher.svelte"
+  import ComponentPropertiesPanel from "components/userInterface/ComponentPropertiesPanel.svelte"
+  import ComponentSelectionList from "components/userInterface/ComponentSelectionList.svelte"
+  import Switcher from "components/common/Switcher.svelte"
   import ConfirmDialog from "components/common/ConfirmDialog.svelte"
   import { last } from "lodash/fp"
   import { AddIcon } from "components/common/Icons"
+  import FrontendNavigatePane from "components/userInterface/FrontendNavigatePane.svelte"
 
   $: instances = $store.appInstances
 
@@ -27,13 +26,10 @@
     }
   })
 
-  let newScreenPicker
+  
   let confirmDeleteDialog
   let componentToDelete = ""
 
-  const newScreen = () => {
-    newScreenPicker.show()
-  }
 
   let settingsView
   const settings = () => {
@@ -45,6 +41,8 @@
     confirmDeleteDialog.show()
   }
 
+  let leftNavSwitcher
+
   const lastPartOfName = c => (c ? last(c.split("/")) : "")
 </script>
 
@@ -52,54 +50,32 @@
 
   <div class="ui-nav">
 
-    <div class="pages-list-container">
-      <div class="nav-header">
-        <span class="navigator-title">Navigate</span>
-        <span class="components-nav-page">Pages</span>
+    <Switcher bind:this={leftNavSwitcher} tabs={["Navigate", "Add"]}>
+      <div slot="0">
+        <FrontendNavigatePane />
       </div>
-
-      <div class="nav-items-container">
-        <PagesList />
-      </div>
-    </div>
-
-    <div class="border-line" />
-
-    <PageLayout layout={$store.pages[$store.currentPageName]} />
-
-    <div class="border-line" />
-
-    <div class="components-list-container">
-      <div class="nav-group-header">
-        <span class="components-nav-header" style="margin-top: 0;">
-          Screens
-        </span>
-        <div>
-          <button on:click={newScreen}>
-            <AddIcon />
-          </button>
-        </div>
-      </div>
-      <div class="nav-items-container">
-        <ComponentsHierarchy screens={$store.screens} />
-      </div>
-    </div>
+      <div slot="1">
+        <ComponentSelectionList toggleTab={leftNavSwitcher.selectTab} />      </div>
+    </Switcher>
+    
 
   </div>
 
   <div class="preview-pane">
+    {#if $store.currentPageName && $store.currentPageName.length > 0}
     <CurrentItemPreview />
+    {/if}
   </div>
 
   {#if $store.currentFrontEndType === 'screen' || $store.currentFrontEndType === 'page'}
     <div class="components-pane">
-      <ComponentsPaneSwitcher />
+      <ComponentPropertiesPanel />
     </div>
   {/if}
 
 </div>
 
-<NewScreen bind:this={newScreenPicker} />
+
 
 <ConfirmDialog
   bind:this={confirmDeleteDialog}
@@ -111,19 +87,6 @@
 <slot />
 
 <style>
-  button {
-    cursor: pointer;
-    outline: none;
-    border: none;
-    border-radius: 5px;
-    width: 20px;
-    padding-bottom: 10px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 0;
-  }
-
   .root {
     display: grid;
     grid-template-columns: 275px 1fr 275px;
@@ -162,43 +125,7 @@
     background-color: var(--white);
   }
 
-  .components-nav-page {
-    font-size: 13px;
-    color: var(--ink);
-    padding-left: 20px;
-    margin-top: 20px;
-    font-weight: 600;
-    opacity: 0.4;
-    letter-spacing: 1px;
-  }
-
-  .components-nav-header {
-    font-size: 13px;
-    color: var(--ink);
-    margin-top: 20px;
-    font-weight: 600;
-    opacity: 0.4;
-    letter-spacing: 1px;
-  }
-
-  .nav-header {
-    display: flex;
-    flex-direction: column;
-    margin-top: 20px;
-  }
-
-  .nav-items-container {
-    padding: 1rem 0rem 0rem 0rem;
-  }
-
-  .nav-group-header {
-    display: flex;
-    padding: 0px 20px 0px 20px;
-    font-size: 0.9rem;
-    font-weight: bold;
-    justify-content: space-between;
-    align-items: center;
-  }
+  
 
   .nav-group-header > div:nth-child(1) {
     padding: 0rem 0.5rem 0rem 0rem;
@@ -207,12 +134,7 @@
     margin-right: 5px;
   }
 
-  .nav-group-header > span:nth-child(3) {
-    margin-left: 5px;
-    vertical-align: bottom;
-    grid-column-start: title;
-    margin-top: auto;
-  }
+
 
   .nav-group-header > div:nth-child(3) {
     vertical-align: bottom;
@@ -225,18 +147,4 @@
     color: var(--primary75);
   }
 
-  .navigator-title {
-    font-size: 18px;
-    color: var(--ink);
-    font-weight: bold;
-    padding: 0 20px 20px 20px;
-  }
-
-  .border-line {
-    border-bottom: 1px solid #d8d8d8;
-  }
-
-  .components-list-container {
-    padding: 20px 0px 0 0;
-  }
 </style>
